requirements/requirements.txt
django==3.2.13
djangorestframework
django-redis
psycopg2
celery
boto3
watchtower
gunicorn
python-decouple
requests
django-smart-selects==1.6.0
django-admin-rangefilter
cerberus
django-extensions
django-postgrespool2==2.0.5
django-crispy-forms==1.14.0
django-model-utils==4.2.0
django-autocomplete-light==3.9.4